tests: move #143 race tests into their own fx3_lifecycle module
Follow the modularization pattern (as with fx3_fuzz) instead of piling the
new tests into the already-large fx3_cmd.c: extract reopen_race_storm and
two_actor_open verbatim into fx3_lifecycle.{c,h}. fx3_cmd.c keeps only the
CLI/console wiring (local_cmds, main dispatch, usage) and gains the include;
the forward declaration is dropped (now in the header). Makefile and README
file-map updated. open_rx888()'s backoff stays in fx3_usb (its module).

No behavior change; builds clean under -Wall -Wextra, check-health passes.
